Refactor to avoid spurious XXXKt classes in api.txt (#6303)
