+++ firebase-dataconnect/src/main/kotlin/com/google/firebase/dataconnect/FirebaseDataConnect.kt
@@ -21,6 +21,7 @@ import com.google.firebase.Firebase
 import com.google.firebase.FirebaseApp
 import com.google.firebase.app
 import com.google.firebase.dataconnect.core.FirebaseDataConnectFactory
+import com.google.firebase.dataconnect.core.LoggerGlobals
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.serialization.DeserializationStrategy
 import kotlinx.serialization.SerializationStrategy
@@ -440,8 +441,4 @@ public fun FirebaseDataConnect.Companion.getInstance(
  * logging, which is especially useful when reporting issues to Google or investigating problems
  * yourself. Setting it to [LogLevel.NONE] will disable all logging.
  */
-public var FirebaseDataConnect.Companion.logLevel: LogLevel
-  get() = com.google.firebase.dataconnect.core.logLevel
-  set(newLogLevel) {
-    com.google.firebase.dataconnect.core.logLevel = newLogLevel
-  }
+public var FirebaseDataConnect.Companion.logLevel: LogLevel by LoggerGlobals::logLevel
