+++ firebase-dataconnect/src/main/kotlin/com/google/firebase/dataconnect/core/QueryRefImpl.kt
@@ -77,24 +77,3 @@ internal class QueryRefImpl<Data, Variables>(
     override fun toString() = "QueryResultImpl(data=$data, ref=$ref)"
   }
 }
-
-internal fun <Data, Variables> QueryRefImpl<Data, Variables>.copy(
-  dataConnect: FirebaseDataConnectInternal = this.dataConnect,
-  operationName: String = this.operationName,
-  variables: Variables = this.variables,
-  dataDeserializer: DeserializationStrategy<Data> = this.dataDeserializer,
-  variablesSerializer: SerializationStrategy<Variables> = this.variablesSerializer,
-  callerSdkType: FirebaseDataConnect.CallerSdkType = this.callerSdkType,
-  variablesSerializersModule: SerializersModule? = this.variablesSerializersModule,
-  dataSerializersModule: SerializersModule? = this.dataSerializersModule,
-) =
-  QueryRefImpl(
-    dataConnect = dataConnect,
-    operationName = operationName,
-    variables = variables,
-    dataDeserializer = dataDeserializer,
-    variablesSerializer = variablesSerializer,
-    callerSdkType = callerSdkType,
-    variablesSerializersModule = variablesSerializersModule,
-    dataSerializersModule = dataSerializersModule,
-  )
