+++ firebase-dataconnect/src/main/kotlin/com/google/firebase/dataconnect/serializers/UUIDSerializer.kt
@@ -16,7 +16,6 @@
 
 package com.google.firebase.dataconnect.serializers
 
-import androidx.annotation.VisibleForTesting
 import java.util.UUID
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.descriptors.PrimitiveKind
@@ -34,22 +33,22 @@ public object UUIDSerializer : KSerializer<UUID> {
     PrimitiveSerialDescriptor("UUID", PrimitiveKind.STRING)
 
   override fun serialize(encoder: Encoder, value: UUID) {
-    val uuidString = serialize(value)
+    val uuidString = UUIDSerializerImpl.serialize(value)
     encoder.encodeString(uuidString)
   }
 
-  @VisibleForTesting
+  override fun deserialize(decoder: Decoder): UUID {
+    val decodedString = decoder.decodeString()
+    return UUIDSerializerImpl.deserialize(decodedString)
+  }
+}
+
+internal object UUIDSerializerImpl {
   internal fun serialize(value: UUID): String {
     // Remove dashes from the UUID since the server will remove them anyways (see cl/629562890).
     return value.toString().replace("-", "")
   }
 
-  override fun deserialize(decoder: Decoder): UUID {
-    val decodedString = decoder.decodeString()
-    return deserialize(decodedString)
-  }
-
-  @VisibleForTesting
   internal fun deserialize(decodedString: String): UUID {
     require(decodedString.length == 32) {
       "invalid UUID string: $decodedString (length=${decodedString.length}, expected=32)"
