+++ firebase-dataconnect/src/main/kotlin/com/google/firebase/dataconnect/util/NullOutputStream.kt
@@ -1,0 +1,24 @@
+/*
+ * Copyright 2024 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.firebase.dataconnect.util
+
+import java.io.OutputStream
+
+internal object NullOutputStream : OutputStream() {
+  override fun write(b: Int) {}
+  override fun write(b: ByteArray?) {}
+  override fun write(b: ByteArray?, off: Int, len: Int) {}
+}
