+++ firebase-dataconnect/src/test/kotlin/com/google/firebase/dataconnect/AnyValueUnitTest.kt
@@ -25,7 +25,7 @@ import com.google.firebase.dataconnect.testutil.anyNumberScalar
 import com.google.firebase.dataconnect.testutil.anyScalar
 import com.google.firebase.dataconnect.testutil.anyStringScalar
 import com.google.firebase.dataconnect.testutil.filterNotNull
-import com.google.firebase.dataconnect.util.encodeToValue
+import com.google.firebase.dataconnect.util.ProtoUtil.encodeToValue
 import io.kotest.assertions.assertSoftly
 import io.kotest.common.ExperimentalKotest
 import io.kotest.matchers.booleans.shouldBeFalse
