+++ firebase-dataconnect/src/test/kotlin/com/google/firebase/dataconnect/ProtoStructDecoderUnitTest.kt
@@ -19,9 +19,9 @@
 package com.google.firebase.dataconnect
 
 import com.google.common.truth.Truth.assertThat
-import com.google.firebase.dataconnect.util.buildStructProto
-import com.google.firebase.dataconnect.util.decodeFromStruct
-import com.google.firebase.dataconnect.util.encodeToStruct
+import com.google.firebase.dataconnect.util.ProtoUtil.buildStructProto
+import com.google.firebase.dataconnect.util.ProtoUtil.decodeFromStruct
+import com.google.firebase.dataconnect.util.ProtoUtil.encodeToStruct
 import com.google.protobuf.Struct
 import com.google.protobuf.Value
 import com.google.protobuf.Value.KindCase
