+++ firebase-dataconnect/src/test/kotlin/com/google/firebase/dataconnect/UtilUnitTest.kt
@@ -17,7 +17,7 @@
 package com.google.firebase.dataconnect
 
 import com.google.common.truth.Truth.assertThat
-import com.google.firebase.dataconnect.util.toAlphaNumericString
+import com.google.firebase.dataconnect.util.AlphanumericStringUtil.toAlphaNumericString
 import org.junit.Test
 
 class UtilUnitTest {
