+++ firebase-dataconnect/src/test/kotlin/com/google/firebase/dataconnect/core/MutationRefImplUnitTest.kt
@@ -21,14 +21,17 @@ import com.google.firebase.dataconnect.DataConnectUntypedData
 import com.google.firebase.dataconnect.DataConnectUntypedVariables
 import com.google.firebase.dataconnect.FirebaseDataConnect.CallerSdkType
 import com.google.firebase.dataconnect.core.DataConnectGrpcClient.OperationResult
+import com.google.firebase.dataconnect.core.Globals.copy
+import com.google.firebase.dataconnect.core.Globals.withDataDeserializer
+import com.google.firebase.dataconnect.core.Globals.withVariablesSerializer
 import com.google.firebase.dataconnect.testutil.callerSdkType
 import com.google.firebase.dataconnect.testutil.dataConnectError
 import com.google.firebase.dataconnect.testutil.filterNotEqual
 import com.google.firebase.dataconnect.testutil.mutationRefImpl
+import com.google.firebase.dataconnect.util.ProtoUtil.buildStructProto
+import com.google.firebase.dataconnect.util.ProtoUtil.encodeToStruct
+import com.google.firebase.dataconnect.util.ProtoUtil.toStructProto
 import com.google.firebase.dataconnect.util.SuspendingLazy
-import com.google.firebase.dataconnect.util.buildStructProto
-import com.google.firebase.dataconnect.util.encodeToStruct
-import com.google.firebase.dataconnect.util.toStructProto
 import com.google.protobuf.Struct
 import io.kotest.assertions.asClue
 import io.kotest.assertions.assertSoftly
