+++ firebase-dataconnect/src/test/kotlin/com/google/firebase/dataconnect/testutil/Arbs.kt
@@ -21,7 +21,7 @@ import com.google.firebase.dataconnect.DataConnectError
 import com.google.firebase.dataconnect.core.DataConnectGrpcClient
 import com.google.firebase.dataconnect.core.MutationRefImpl
 import com.google.firebase.dataconnect.core.QueryRefImpl
-import com.google.firebase.dataconnect.util.toStructProto
+import com.google.firebase.dataconnect.util.ProtoUtil.toStructProto
 import io.kotest.property.Arb
 import io.kotest.property.arbitrary.Codepoint
 import io.kotest.property.arbitrary.alphanumeric
